MediaCache: replaced 'binary-serialize' (C++) with 'node-pack'
Removed the native C++ addon which used cereal to serialize data into
binary format and replaced it with 'node-pack' which also serializes
data into binary format.

The reason why I did this is for the sake of disk space. node-pack
produces way smaller binary files than cereal does.

This should reduce the cache directory size about ~40%
It is even smaller now than it is in the C++ version of Music Console :)