Common: bring back an improved version of the GNU/Readline addon
Since I couldn't find a satisfying prompt for Node I tried my best
to improve my GNU/Readline addon.

I managed to make signal handling work and at the same time I
added a shell-like Ctrl+C functionality which cancels the current
prompt instance and starts a new one in the next line.

GNU/Readline is in many ways better than every prompt for Node I
tried. It has proper Unicode and wide character support. It has
a neat history feature. And the best of it, it BLOCKS the I/O
of Node so its execution is always synchronized :)

There is still a issue with Ctrl+D which is available in KDE/Konsole
and maybe other terminal emulators. The shortcut cancels readline,
but only when the line is empty - if not it does nothing.
This is not that much of a problem for this project since an empty
input is ignored anyway and jumps back to the prompt.