lib/kbhit: implemented kbhit (native addon)
The kbhit function waits a specific time for user
response to break out of a(n) (infinite) loop.
Break key == Carriage Return [ONLY]

This makes it possible to make user breakable
infinite loops or break out of a loop early.

This is required for the following features which
will be implemented in the future:
  - Playlist
  - Shuffle
You can think of it now why this is required.