[SPECIAL] (read commit message)
I learned a new keyword today -> 'let' and also I learned
the hard way why it is bad to extend JavaScript globals
using '.prototype.myextension'. Related to this two findings
I made a lot of changes which affects the performance in a
positive manner.

I changed every single loop in the program to use either
'let' or 'const' depending on if the value needs to be changed
or not.

I learned the hard way what happens when you extend globals
in JavaScript and than use (for...in) and (for...of) loops.
The loop contains every customization of the globals.
I removed all global extensions and wrote a new Node module
which resides in 'utils' for now.