PlaylistParser: playlist file validator implemented
In preparation to the playlist parser, a validator got implemented.
The validator checks every line individually if it matches the
syntax pattern which is 'T:(T):STRING (PLAYER)'.

The validator has the ability to report the position of the
problematic character/phrase which mismatches the syntax pattern.
This feature is experimental, the position is off at some places.

As an extend the validator also minifies the lines to remove
any unnecessary spaces (of course only outside of strings).

The difference from the new version compared to the old version
implemented in C++ is that, that this validator will no longer
attempt to fix user errors and ignores falsy entries, but rather
stops and reports the problem. The user now must fix the playlist
before going any futher.

TODO: Now its time to implement an easy parser, since there
is already a validator, the parser code can stay simple and
doesn't need to perform any kind of unnecessary checks :)