[Win] Re-added Windows support (with limited functionality)
Music Console is now functional under Windows with very
limited functionality. I will not support this OS due to
the lack of proper Unicode and Color support in the Terminal.

There are many bugs and the Tag Reader is disabled. The
media player controller is not very stable and there
is also no signal handling (blame Microsoft not me).

Developing command line apps for Windows is not fun at all.